requirements.txt
open3d==0.17.0
ipdb==0.13.13
timm==0.6.7
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchinfo==1.8.0
torchmetrics==0.11.4
torchvision==0.15.2+cu118
opencv-contrib-python==4.8.0.74
Pillow==9.3.0
